requirements.txt
biopython>=1.82
pyhmmer>=0.10.0
numpy>=1.24.0
jinja2>=3.1.0
